refactor: remove legacy Android UI and adopt Compose Multiplatform
