+++ gradle/libs.versions.toml
@@ -15,6 +15,7 @@ androidx-splashscreen = "1.2.0"
 
 # Compose
 bomCompose = "2026.02.00"
+composeMultiplatform = "1.11.0"
 
 # Navigation
 decompose = "3.5.0"
@@ -117,6 +118,9 @@ ksp = { id = "com.google.devtools.ksp", version.ref = "ksp" }
 
 # Navigation
 jetbrains-kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlinSerialization"}
+
+# Compose Multiplatform
+compose-multiplatform = { id = "org.jetbrains.compose", version.ref = "composeMultiplatform" }
 
 #DI
 kotlin-multiplatform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
