--- presentation/src/androidMain/kotlin/com/mindeck/presentation/ui/components/buttons/CustomButton.kt
@@ -1,40 +1,0 @@
-package com.mindeck.presentation.ui.components.buttons
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-
-@Composable
-fun CustomButton(
-    text: String,
-    color: Color = MaterialTheme.colorScheme.primary,
-    textColor: Color = MaterialTheme.colorScheme.onPrimary,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier =
-            modifier
-                .clip(MaterialTheme.shapes.medium)
-                .background(
-                    color = color,
-                    shape = MaterialTheme.shapes.medium,
-                ).clickable {
-                    onClick()
-                },
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(
-            text = text,
-            style = MaterialTheme.typography.titleMedium,
-            color = textColor,
-        )
-    }
-}
