--- presentation/src/androidMain/kotlin/com/mindeck/presentation/ui/navigation/Navigator.kt
@@ -1,12 +1,0 @@
-package com.mindeck.presentation.ui.navigation
-
-import androidx.compose.runtime.staticCompositionLocalOf
-
-val LocalNavigator = staticCompositionLocalOf<Navigator> {
-    error("LocalNavigator not provided")
-}
-
-interface Navigator {
-    fun push(route: NavigationRoute)
-    fun pop()
-}
