--- presentation/src/androidMain/kotlin/com/mindeck/presentation/ui/screens/CardStudyScreen.kt
@@ -1,331 +1,0 @@
-package com.mindeck.presentation.ui.screens
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.foundation.layout.systemBarsPadding
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.mindeck.domain.models.Card
-import com.mindeck.domain.models.CardState
-import com.mindeck.domain.models.ReviewButton
-import com.mindeck.presentation.R
-import com.mindeck.presentation.state.ModalState
-import com.mindeck.presentation.state.UiState
-import com.mindeck.presentation.ui.components.dropdown.AppDropdownMenu
-import com.mindeck.presentation.ui.components.dropdown.AppDropdownMenuItem
-import com.mindeck.presentation.ui.components.repeatOptions.RepeatOptionsButton
-import com.mindeck.presentation.ui.components.topBar.AppTopBar
-import com.mindeck.presentation.ui.model.RepeatOptionData
-import com.mindeck.presentation.ui.navigation.LocalNavigator
-import com.mindeck.presentation.ui.theme.MindeckTheme
-import com.mindeck.presentation.ui.theme.repeat_option_again_dark
-import com.mindeck.presentation.ui.theme.repeat_option_again_light
-import com.mindeck.presentation.ui.theme.repeat_option_easy_dark
-import com.mindeck.presentation.ui.theme.repeat_option_easy_light
-import com.mindeck.presentation.ui.theme.repeat_option_hard_dark
-import com.mindeck.presentation.ui.theme.repeat_option_hard_light
-import com.mindeck.presentation.ui.theme.repeat_option_medium_dark
-import com.mindeck.presentation.ui.theme.repeat_option_medium_light
-import com.mindeck.presentation.viewmodel.card.CardStudyViewModel
-import com.mohamedrejeb.richeditor.annotation.ExperimentalRichTextApi
-import com.mohamedrejeb.richeditor.model.rememberRichTextState
-import com.mohamedrejeb.richeditor.ui.material3.RichText
-import org.koin.compose.viewmodel.koinViewModel
-
-@Composable
-fun CardStudyScreen(
-    cardId: Int? = null,
-    modifier: Modifier = Modifier,
-) {
-    val navigator = LocalNavigator.current
-
-    val viewModel = koinViewModel<CardStudyViewModel>()
-    val modalState by viewModel.modalState.collectAsStateWithLifecycle()
-    val cardsState by viewModel.cardsState.collectAsStateWithLifecycle()
-    val reviewLabels by viewModel.reviewLabels.collectAsStateWithLifecycle()
-
-    LaunchedEffect(Unit) {
-        if (cardId != null) {
-            viewModel.loadCardById(cardId)
-        } else {
-            viewModel.loadCardRepetition()
-        }
-    }
-
-    CardStudyScreenContent(
-        modalState = modalState,
-        cardsForRepetitionState = cardsState,
-        reviewLabels = reviewLabels,
-        actions =
-            CardStudyScreenActions(
-                onNavigateBack = navigator::pop,
-                onShowDropdownMenu = viewModel::showDropdownMenu,
-                onHideModal = viewModel::hideModal,
-                onReviewCard = viewModel::reviewCard,
-            ),
-        modifier = modifier,
-    )
-}
-
-@Composable
-private fun formatInterval(millis: Long): String {
-    val minutes =
-        java.util.concurrent.TimeUnit.MILLISECONDS
-            .toMinutes(millis)
-    val hours =
-        java.util.concurrent.TimeUnit.MILLISECONDS
-            .toHours(millis)
-    return when {
-        millis <
-            java.util.concurrent.TimeUnit.MINUTES
-                .toMillis(1)
-        ->
-            stringResource(
-                R.string.review_interval_seconds,
-                java.util.concurrent.TimeUnit.MILLISECONDS
-                    .toSeconds(millis),
-            )
-        hours < 1 ->
-            stringResource(R.string.review_interval_minutes, minutes)
-        else ->
-            stringResource(
-                R.string.review_interval_days,
-                kotlin.math
-                    .ceil(
-                        millis.toDouble() /
-                            java.util.concurrent.TimeUnit.DAYS
-                                .toMillis(1),
-                    ).toInt(),
-            )
-    }
-}
-
-@OptIn(ExperimentalRichTextApi::class)
-@Composable
-internal fun CardStudyScreenContent(
-    modalState: ModalState,
-    cardsForRepetitionState: UiState<List<Card>>,
-    reviewLabels: Map<ReviewButton, Long>,
-    actions: CardStudyScreenActions,
-    modifier: Modifier = Modifier,
-) {
-    val isDark = isSystemInDarkTheme()
-    val scrollState = rememberScrollState()
-
-    val questionState = rememberRichTextState()
-    val answerState = rememberRichTextState()
-
-    Scaffold(
-        modifier =
-            modifier
-                .fillMaxSize()
-                .systemBarsPadding(),
-        containerColor = MaterialTheme.colorScheme.background,
-        topBar = {
-            AppTopBar(
-                showMenuButton = cardsForRepetitionState is UiState.Success,
-                onBackClick = actions.onNavigateBack,
-                onMenuClick = actions.onShowDropdownMenu,
-                modifier = Modifier.padding(horizontal = MindeckTheme.dimensions.paddingMd),
-            )
-        },
-        content = { padding ->
-            when (cardsForRepetitionState) {
-                is UiState.Success -> {
-                    val cards = cardsForRepetitionState.data
-                    val currentCard = cards.firstOrNull()
-
-                    LaunchedEffect(currentCard?.cardId) { scrollState.scrollTo(0) }
-
-                    LaunchedEffect(currentCard?.cardQuestion) {
-                        currentCard?.cardQuestion?.let { questionState.setHtml(it) }
-                    }
-
-                    LaunchedEffect(currentCard?.cardAnswer) {
-                        currentCard?.cardAnswer?.let { answerState.setHtml(it) }
-                    }
-
-                    Column(
-                        modifier =
-                            Modifier
-                                .padding(padding)
-                                .padding(horizontal = MindeckTheme.dimensions.paddingMd)
-                                .statusBarsPadding()
-                                .verticalScroll(state = scrollState),
-                        horizontalAlignment = Alignment.CenterHorizontally,
-                    ) {
-                        if (currentCard == null) {
-                            Spacer(modifier = Modifier.height(MindeckTheme.dimensions.dp30))
-                            Text(
-                                text = stringResource(R.string.study_session_complete_text),
-                                style = MaterialTheme.typography.titleLarge,
-                                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                textAlign = TextAlign.Center,
-                            )
-                        } else {
-                            RichText(
-                                state = questionState,
-                                style = MaterialTheme.typography.titleLarge,
-                            )
-                            Spacer(modifier = Modifier.height(MindeckTheme.dimensions.spacerSm))
-                            Box(
-                                modifier =
-                                    Modifier
-                                        .fillMaxWidth()
-                                        .clip(MaterialTheme.shapes.large)
-                                        .background(MaterialTheme.colorScheme.primary)
-                                        .height(MindeckTheme.dimensions.dp1)
-                                        .padding(horizontal = MindeckTheme.dimensions.paddingMd),
-                            )
-                            Spacer(modifier = Modifier.height(MindeckTheme.dimensions.spacerSm))
-                            RichText(
-                                state = answerState,
-                                style = MaterialTheme.typography.titleLarge,
-                            )
-                        }
-                    }
-
-                    if (currentCard != null) {
-                        val showAgain =
-                            currentCard.cardState == CardState.LAPSE ||
-                                currentCard.cardState == CardState.REVIEW ||
-                                (currentCard.cardState == CardState.LEARNING && currentCard.learningStep > 0)
-                        RepeatButtons(
-                            options =
-                                listOfNotNull(
-                                    if (showAgain) {
-                                        RepeatOptionData(
-                                            title = stringResource(R.string.repeat_option_title_again_text),
-                                            time = reviewLabels[ReviewButton.AGAIN]?.let { formatInterval(it) }.orEmpty(),
-                                            color = if (isDark) repeat_option_again_dark else repeat_option_again_light,
-                                            onClick = { actions.onReviewCard(currentCard, ReviewButton.AGAIN) },
-                                        )
-                                    } else {
-                                        null
-                                    },
-                                    RepeatOptionData(
-                                        title = stringResource(R.string.repeat_option_title_hard_text),
-                                        time = reviewLabels[ReviewButton.HARD]?.let { formatInterval(it) }.orEmpty(),
-                                        color = if (isDark) repeat_option_hard_dark else repeat_option_hard_light,
-                                        onClick = { actions.onReviewCard(currentCard, ReviewButton.HARD) },
-                                    ),
-                                    RepeatOptionData(
-                                        title = stringResource(R.string.repeat_option_title_good_text),
-                                        time = reviewLabels[ReviewButton.GOOD]?.let { formatInterval(it) }.orEmpty(),
-                                        color = if (isDark) repeat_option_medium_dark else repeat_option_medium_light,
-                                        onClick = { actions.onReviewCard(currentCard, ReviewButton.GOOD) },
-                                    ),
-                                    RepeatOptionData(
-                                        title = stringResource(R.string.repeat_option_title_easy_text),
-                                        time = reviewLabels[ReviewButton.EASY]?.let { formatInterval(it) }.orEmpty(),
-                                        color = if (isDark) repeat_option_easy_dark else repeat_option_easy_light,
-                                        onClick = { actions.onReviewCard(currentCard, ReviewButton.EASY) },
-                                    ),
-                                ),
-                        )
-                    }
-
-                    AppDropdownMenu(
-                        padding = padding,
-                        isExpanded = modalState is ModalState.DropdownMenu,
-                        onDismiss = actions.onHideModal,
-                    ) {
-                        AppDropdownMenuItem(
-                            text = stringResource(R.string.dropdown_menu_data_previous_card),
-                            onClick = {
-                                actions.onHideModal()
-                            },
-                        )
-                    }
-                }
-
-                UiState.Loading -> {
-                    Column(
-                        modifier =
-                            Modifier
-                                .fillMaxWidth()
-                                .wrapContentSize(Alignment.Center),
-                    ) {
-                        Spacer(modifier = Modifier.height(MindeckTheme.dimensions.spacerMd))
-                        CircularProgressIndicator(
-                            color = MaterialTheme.colorScheme.primary,
-                            strokeWidth = MindeckTheme.dimensions.dp2,
-                        )
-                    }
-                }
-
-                is UiState.Error -> {
-                    Column(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalAlignment = Alignment.CenterHorizontally,
-                        verticalArrangement = Arrangement.spacedBy(MindeckTheme.dimensions.dp16),
-                    ) {
-                        Spacer(modifier = Modifier.height(MindeckTheme.dimensions.dp80))
-                        Text(
-                            stringResource(R.string.error_get_card_for_study),
-                            modifier = Modifier.fillMaxWidth(),
-                            textAlign = TextAlign.Center,
-                            style = MaterialTheme.typography.bodyLarge,
-                            color = MaterialTheme.colorScheme.error,
-                        )
-                    }
-                }
-
-                UiState.Idle -> Unit
-            }
-        },
-    )
-}
-
-@Composable
-private fun RepeatButtons(options: List<RepeatOptionData>) {
-    Row(
-        horizontalArrangement = Arrangement.SpaceAround,
-        verticalAlignment = Alignment.Bottom,
-        modifier =
-            Modifier
-                .fillMaxSize(),
-    ) {
-        options.forEach {
-            RepeatOptionsButton(
-                buttonColor = it.color,
-                label = it.title,
-                time = it.time,
-                onClick = it.onClick,
-                textStyle = MaterialTheme.typography.bodyMedium,
-                modifier = Modifier.weight(1f),
-            )
-        }
-    }
-}
-
-data class CardStudyScreenActions(
-    val onNavigateBack: () -> Unit,
-    val onShowDropdownMenu: () -> Unit,
-    val onHideModal: () -> Unit,
-    val onReviewCard: (Card, ReviewButton) -> Unit,
-)
