--- presentation/src/androidMain/kotlin/com/mindeck/presentation/ui/theme/Colors.kt
@@ -1,75 +1,0 @@
-package com.mindeck.presentation.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val md_light_primary = Color(0xFF3F5F90)
-val md_light_onPrimary = Color(0xFFFFFFFF)
-val md_light_primaryContainer = Color(0xFFD5E3FF)
-val md_light_onPrimaryContainer = Color(0xFF254777)
-
-val md_light_secondary = Color(0xFF555F71)
-val md_light_onSecondary = Color(0xFFFFFFFF)
-val md_light_secondaryContainer = Color(0xFFD9E3F8)
-val md_light_onSecondaryContainer = Color(0xFF3E4758)
-
-val md_light_tertiary = Color(0xFF6F5675)
-val md_light_onTertiary = Color(0xFFFFFFFF)
-val md_light_tertiaryContainer = Color(0xFFF8D8FE)
-val md_light_onTertiaryContainer = Color(0xFF563E5D)
-
-val md_light_error = Color(0xFFBA1A1A)
-val md_light_onError = Color(0xFFFFFFFF)
-val md_light_errorContainer = Color(0xFFFFDAD6)
-val md_light_onErrorContainer = Color(0xFF93000A)
-
-val md_light_background = Color(0xFFF9F9FF)
-val md_light_onBackground = Color(0xFF191C20)
-val md_light_surface = Color(0xFFF9F9FF)
-val md_light_onSurface = Color(0xFF191C20)
-val md_light_surfaceVariant = Color(0xFFE0E2EC)
-val md_light_onSurfaceVariant = Color(0xFF43474E)
-
-val md_light_outline = Color(0xFF74777F)
-val md_light_outlineVariant = Color(0xFFC4C6CF)
-val md_light_scrim = Color(0xFF000000)
-
-val md_dark_primary = Color(0xFFA8C8FF)
-val md_dark_onPrimary = Color(0xFF06305F)
-val md_dark_primaryContainer = Color(0xFF254777)
-val md_dark_onPrimaryContainer = Color(0xFFD5E3FF)
-
-val md_dark_secondary = Color(0xFFBDC7DC)
-val md_dark_onSecondary = Color(0xFF273141)
-val md_dark_secondaryContainer = Color(0xFF3E4758)
-val md_dark_onSecondaryContainer = Color(0xFFD9E3F8)
-
-val md_dark_tertiary = Color(0xFFDBBCE1)
-val md_dark_onTertiary = Color(0xFF3E2845)
-val md_dark_tertiaryContainer = Color(0xFF563E5D)
-val md_dark_onTertiaryContainer = Color(0xFFF8D8FE)
-
-val md_dark_error = Color(0xFFFFB4AB)
-val md_dark_onError = Color(0xFF690005)
-val md_dark_errorContainer = Color(0xFF93000A)
-val md_dark_onErrorContainer = Color(0xFFFFDAD6)
-
-val md_dark_background = Color(0xFF111318)
-val md_dark_onBackground = Color(0xFFE1E2E9)
-val md_dark_surface = Color(0xFF111318)
-val md_dark_onSurface = Color(0xFFE1E2E9)
-val md_dark_surfaceVariant = Color(0xFF43474E)
-val md_dark_onSurfaceVariant = Color(0xFFC4C6CF)
-
-val md_dark_outline = Color(0xFF8E9099)
-val md_dark_outlineVariant = Color(0xFF43474E)
-val md_dark_scrim = Color(0xFF000000)
-
-val repeat_option_again_light = Color(0xFFD9E3F8)
-val repeat_option_easy_light = Color(0xFFCEF0DA)
-val repeat_option_medium_light = Color(0xFFFFF8C6)
-val repeat_option_hard_light = Color(0xFFFFDAD6)
-
-val repeat_option_again_dark = Color(0xFF3E4758)
-val repeat_option_easy_dark = Color(0xFF003920)
-val repeat_option_medium_dark = Color(0xFF362B00)
-val repeat_option_hard_dark = Color(0xFF93000A)
