--- presentation/src/androidMain/kotlin/com/mindeck/presentation/viewmodel/card/CardStudyViewModel.kt
@@ -1,136 +1,0 @@
-package com.mindeck.presentation.viewmodel.card
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.mindeck.domain.exception.DomainError
-import com.mindeck.domain.models.Card
-import com.mindeck.domain.models.CardState
-import com.mindeck.domain.models.ReviewButton
-import com.mindeck.domain.usecases.card.command.UpdateCardReviewUseCase
-import com.mindeck.domain.usecases.card.query.GetCardByIdUseCase
-import com.mindeck.domain.usecases.card.query.GetCardsRepetitionUseCase
-import com.mindeck.presentation.R
-import com.mindeck.presentation.state.ModalState
-import com.mindeck.presentation.state.UiState
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.flow.update
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.sync.Mutex
-import kotlinx.coroutines.sync.withLock
-
-internal class CardStudyViewModel(
-    private val getCardsRepetitionUseCase: GetCardsRepetitionUseCase,
-    private val getCardByIdUseCase: GetCardByIdUseCase,
-    private val updateCardReviewUseCase: UpdateCardReviewUseCase,
-) : ViewModel() {
-    private val _modalState = MutableStateFlow<ModalState>(ModalState.None)
-    val modalState: StateFlow<ModalState> = _modalState.asStateFlow()
-
-    private val _cardsState = MutableStateFlow<UiState<List<Card>>>(UiState.Idle)
-    val cardsState: StateFlow<UiState<List<Card>>> = _cardsState.asStateFlow()
-
-    private val _reviewLabels = MutableStateFlow<Map<ReviewButton, Long>>(emptyMap())
-    val reviewLabels: StateFlow<Map<ReviewButton, Long>> = _reviewLabels.asStateFlow()
-
-    private val sessionQueue = ArrayDeque<Card>()
-    private val sessionMutex = Mutex()
-    private val reviewButtonEntries = ReviewButton.entries
-
-    fun loadCardRepetition() {
-        viewModelScope.launch {
-            _cardsState.update { UiState.Loading }
-            try {
-                val cards = getCardsRepetitionUseCase().firstOrNull()
-                if (cards == null) {
-                    _cardsState.update { UiState.Error(R.string.error_get_card_for_study) }
-                    return@launch
-                }
-                sessionMutex.withLock {
-                    sessionQueue.clear()
-                    sessionQueue.addAll(cards)
-                    _cardsState.update { UiState.Success(sessionQueue.toList()) }
-                    updateReviewLabels(sessionQueue.firstOrNull())
-                }
-            } catch (e: DomainError.DatabaseError) {
-                _cardsState.update { UiState.Error(R.string.error_get_card_for_study) }
-            } catch (e: Exception) {
-                _cardsState.update { UiState.Error(R.string.error_something_went_wrong) }
-            }
-        }
-    }
-
-    fun loadCardById(cardId: Int) {
-        viewModelScope.launch {
-            _cardsState.update { UiState.Loading }
-            try {
-                val card = getCardByIdUseCase(cardId = cardId).firstOrNull()
-                sessionMutex.withLock {
-                    sessionQueue.clear()
-                    if (card != null) {
-                        sessionQueue.add(card)
-                        _cardsState.update { UiState.Success(sessionQueue.toList()) }
-                        updateReviewLabels(card)
-                    } else {
-                        _cardsState.update { UiState.Error(R.string.error_failed_to_load_card) }
-                    }
-                }
-            } catch (e: DomainError.DatabaseError) {
-                _cardsState.update { UiState.Error(R.string.error_failed_to_load_card) }
-            } catch (e: Exception) {
-                _cardsState.update { UiState.Error(R.string.error_something_went_wrong) }
-            }
-        }
-    }
-
-    fun reviewCard(
-        card: Card,
-        button: ReviewButton,
-    ) {
-        viewModelScope.launch {
-            sessionMutex.withLock {
-                try {
-                    val updatedCard = updateCardReviewUseCase(card, button)
-
-                    sessionQueue.removeFirstOrNull()
-
-                    if (updatedCard.cardState == CardState.LEARNING || updatedCard.cardState == CardState.LAPSE) {
-                        sessionQueue.addLast(updatedCard)
-                    }
-
-                    _cardsState.update { UiState.Success(sessionQueue.toList()) }
-                    updateReviewLabels(sessionQueue.firstOrNull())
-                } catch (e: Exception) {
-                    _cardsState.update { UiState.Error(R.string.error_something_went_wrong) }
-                }
-            }
-        }
-    }
-
-    private fun updateReviewLabels(card: Card?) {
-        if (card == null) {
-            _reviewLabels.update { emptyMap() }
-            return
-        }
-        _reviewLabels.update {
-            reviewButtonEntries.associateWith { button ->
-                updateCardReviewUseCase.previewNextInterval(card, button)
-            }
-        }
-    }
-
-    fun showDropdownMenu() {
-        _modalState.update { ModalState.DropdownMenu }
-    }
-
-    fun hideModal() {
-        _modalState.update { ModalState.None }
-    }
-
-    override fun onCleared() {
-        super.onCleared()
-        sessionQueue.clear()
-    }
-}
