--- presentation/src/androidMain/kotlin/com/mindeck/presentation/viewmodel/card/CardUiEvent.kt
@@ -1,7 +1,0 @@
-package com.mindeck.presentation.viewmodel.card
-
-sealed interface CardUiEvent {
-    data class DeletionSuccessful(
-        val cardName: String,
-    ) : CardUiEvent
-}
