--- presentation/src/androidMain/kotlin/com/mindeck/presentation/viewmodel/card/CreationCardNavigationEvent.kt
@@ -1,9 +1,0 @@
-package com.mindeck.presentation.viewmodel.card
-
-import androidx.annotation.StringRes
-
-sealed interface CreationCardNavigationEvent {
-    data class ShowToast(
-        @StringRes val messageRes: Int,
-    ) : CreationCardNavigationEvent
-}
