--- presentation/src/androidMain/kotlin/com/mindeck/presentation/viewmodel/deck/DeckScreenData.kt
@@ -1,9 +1,0 @@
-package com.mindeck.presentation.viewmodel.deck
-
-import com.mindeck.domain.models.Card
-import com.mindeck.domain.models.Deck
-
-data class DeckScreenData(
-    val deck: Deck,
-    val cards: List<Card>,
-)
