+++ presentation/src/commonMain/kotlin/com/mindeck/presentation/ui/screen/SecondScreen.kt
@@ -1,18 +1,15 @@
-package com.mindeck.presentation.screen
+package com.mindeck.presentation.ui.screen
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.mindeck.presentation.ui.navigation.Config
 import com.mindeck.presentation.ui.navigation.LocalRootComponent
 
 @Composable
-fun MainScreen(modifier: Modifier = Modifier) {
+fun SecondScreen() {
     val root = LocalRootComponent.current
 
     Column(
@@ -20,9 +17,9 @@ fun MainScreen(modifier: Modifier = Modifier) {
         verticalArrangement = Arrangement.Center,
         modifier = Modifier.fillMaxSize()
     ) {
-        Text(text = "Main Screen")
-        Button(onClick = { root.push(Config.Second) }) {
-            Text(text = "Go to Second")
+        Text(text = "Second Screen")
+        Button(onClick = { root.pop() }) {
+            Text(text = "Go back")
         }
     }
 }
